+++ src/main/java/gift/global/config/WebConfig.java
@@ -1,6 +1,5 @@
 package gift.global.config;
 
-import gift.global.util.LoginMemberArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
